[12.0][MIG] - base_sms_client migration to 12.0
