+++ base_sms_client/__init__.py
@@ -1,7 +1,2 @@
-# -*- coding: utf-8 -*-
-#    Author: Sébastien BEAU <sebastien.beau@akretion.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-
 from . import wizard
 from . import models
