+++ base_sms_client/__manifest__.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Copyright (C) 2015 Sébastien BEAU <sebastien.beau@akretion.com>
 # © 2011 SYLEAM (<http://syleam.fr/>)
 # © 2013 Julius Network Solutions SARL <contact@julius.fr>
@@ -7,7 +6,7 @@
 
 {
     "name": "Base Sms Client",
-    "version": "10.0.1.0.0",
+    "version": "12.0.1.0.0",
     'license': 'AGPL-3',
     "depends": ['mail',
                 'base_phone',
@@ -23,7 +22,7 @@
         'images/send_sms.jpeg'
     ],
     "summary": "Sending SMSs very easily, individually or collectively.",
-    "website": "http://julius.fr",
+    "website": "https://github.com/OCA/connector-telephony",
     "category": "Phone",
     "data": [
         "security/groups.xml",
@@ -35,5 +34,5 @@
         "wizard/mass_sms_view.xml",
         "views/smstemplate_view.xml"
     ],
-    "installable": False,
+    "installable": True,
 }
