+++ base_sms_client/data/sms_gateway_data.xml
@@ -6,10 +6,11 @@
             <field name="interval_type">minutes</field>
             <field name="numbercall">-1</field>
             <field name="doall">False</field>
-            <field name="model">sms.gateway</field>
-            <field name="function">_run_send_sms</field>
-            <field name="args">()</field>
+            <field name="model_id" ref="base_sms_client.model_sms_gateway"/>
+            <field name="state">code</field>
+            <field name="code">model._run_send_sms()</field>
             <field name="active">False</field>
+            <field name="user_id" ref="base.user_root" />
         </record>
 
 </odoo>
