+++ base_sms_client/models/__init__.py
@@ -1,8 +1,4 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2013 Julius Network Solutions SARL <contact@julius.fr>
-#    Author: Sébastien BEAU <sebastien.beau@akretion.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
+from . import sms_abstract
 from . import sms_gateway
 from . import server_action
 from . import sms_template
