+++ base_sms_client/readme/CONFIGURATION.rst
@@ -1,0 +1,4 @@
+To configure this module, you need to:
+
+* Configure Keychain key
+* Configure your provider account
