+++ base_sms_client/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Valentin Chemiere <valentin.chemiere@akretion.com>
+* Yvan Party <yvan@julius.fr>
+* Bejaoui Souheil <souheil.bejaoui@acsone.eu>
