+++ base_sms_client/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module provides base function to send sms. It is not meant to work alone and needs a least one provider module.
