+++ base_sms_client/readme/INSTALL.rst
@@ -1,0 +1,3 @@
+To install this module, you need to:
+
+* Install a provider module after
