+++ base_sms_client/readme/USAGE.rst
@@ -1,0 +1,5 @@
+To use this module, you need to:
+
+ * go to partner menu
+ * select one or multiple partner
+ * go to more and will found some wizards to send sms
