+++ setup/base_sms_client/odoo/addons/base_sms_client
@@ -1,0 +1,1 @@
+../../../../base_sms_client
